Moved `HoodieUnsafeRDDUtils` under `spark.sql` package to get access to package-private API
